+++ price-feeder/oracle/provider/bitget_test.go
@@ -98,7 +98,7 @@ func TestBitgetProvider_GetTickerPrices(t *testing.T) {
 
 	t.Run("invalid_request_invalid_ticker", func(t *testing.T) {
 		prices, err := p.GetTickerPrices(types.CurrencyPair{Base: "FOO", Quote: "BAR"})
-		require.EqualError(t, err, "bitget failed to get ticker price for FOOBAR")
+		require.EqualError(t, err, "bitget has no ticker data for requested pairs: [FOOBAR]")
 		require.Nil(t, prices)
 	})
 }
@@ -139,7 +139,7 @@ func TestBitgetProvider_GetCandlePrices(t *testing.T) {
 
 	t.Run("invalid_request_invalid_candle", func(t *testing.T) {
 		prices, err := p.GetCandlePrices(types.CurrencyPair{Base: "FOO", Quote: "BAR"})
-		require.EqualError(t, err, "failed to get candles price for FOOBAR")
+		require.EqualError(t, err, "bitget has no candle data for requested pairs: [FOOBAR]")
 		require.Nil(t, prices)
 	})
 }